pivotroots_backend/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
PyJWT==2.8.0
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.3